requirements.txt
craft-text-detector
torchvision==0.13
torch
transformers
Pillow
